switch int to uint16_t

+++ cuda/nuc_kernel.h
@@ -1,7 +1,7 @@
 #ifndef KERNEL_H_INCLUDED
 #define KERNEL_H_INCLUDED
-
+#include <stdint.h>
 /* Prototypes for the functions */
-void nucCaller(int *out, float *gain, float *offset, int n);
+void nucCaller(uint16_t *out, float *gain, float *offset, int n);
 
 #endif
